テストをtestファイルに外だし
